+++ examples/misc_coloured.rs
@@ -1,7 +1,7 @@
 use kdam::prelude::*;
 
 fn main() {
-    let mut pb = tqdm!(total = 100, colour = "#a485ca".to_owned());
+    let mut pb = tqdm!(total = 100, colour = "#a485ca");
 
     for _ in 0..100 {
         std::thread::sleep(std::time::Duration::from_secs_f32(0.01));