+++ src/macros.rs
@@ -1,7 +1,8 @@
-/// [tqdm](https://github.com/tqdm/tqdm) like macro for constructing `kdam::Bar` if iterable is not given else `kdam::BarIterator`.
+/// [tqdm](https://github.com/tqdm/tqdm) like macro for constructing `kdam::BarIterator` if iterable is given else `kdam::Bar`.
 ///
-/// See available all [fields](https://docs.rs/kdam/latest/kdam/struct.Bar.html#fields).
-/// 
+/// This macro use `kdam::Bar::Builder` for creating `kdam::Bar`.
+/// See all available [methods](https://docs.rs/kdam/latest/kdam/struct.BarBuilder.html).
+///
 /// # Examples
 ///
 /// ```rust
@@ -9,37 +10,23 @@
 ///
 /// tqdm!();
 /// tqdm!(total = 100);
-/// tqdm!(total = 100, mininterval = 0.0, colour = "green".to_string());
+/// tqdm!(total = 100, mininterval = 0.0, colour = "green");
 /// tqdm!(0..100);
-/// tqdm!(0..100, desc = "0 to 99".to_string());
+/// tqdm!(0..100, desc = "0 to 99");
 /// tqdm!(["a", "b", "c", "d"].iter());
 /// ```
 #[macro_export]
 macro_rules! tqdm {
-    ($($struct_field: ident = $value: expr),*) => {
-        {
-            let mut pb = kdam::Bar::default();
-            $(
-                pb.$struct_field = $value;
-            )*
-            pb
-        }
+    ($($setter_method: ident = $value: expr),*) => {
+        kdam::BarBuilder::default()$(.$setter_method($value))*.build()
     };
 
     ($iterable: expr) => {
-        {
-            kdam::BarIterator::new_with_bar($iterable, kdam::Bar::default())
-        }
+        kdam::BarIterator::new_with_bar($iterable, kdam::Bar::default())
     };
 
-    ($iterable: expr, $($struct_field: ident = $value: expr),*) => {
-        {
-            let mut pb = kdam::Bar::default();
-            $(
-                pb.$struct_field = $value;
-            )*
-            kdam::BarIterator::new_with_bar($iterable, pb)
-        }
+    ($iterable: expr, $($setter_method: ident = $value: expr),*) => {
+        kdam::BarIterator::new_with_bar($iterable, kdam::BarBuilder::default()$(.$setter_method($value))*.build())
     };
 }
 
@@ -51,7 +38,7 @@ macro_rules! tqdm {
 ///
 /// ```rust
 /// use kdam::prelude::*;
-/// 
+///
 /// eprint_at!(1, "1 + 1 = {}", 2);
 /// ```
 #[macro_export]